move plugins from pkg to lib/migration/plugins/
