+++ lib/migration/plugins/init.go
@@ -16,6 +16,6 @@ package plugins
 
 // import all plugin packages here to register plugins
 import (
-	_ "github.com/vmware/vic/pkg/version/plugin2"
-	_ "github.com/vmware/vic/pkg/version/plugin3"
+	_ "github.com/vmware/vic/lib/migration/plugins/plugin1"
+	_ "github.com/vmware/vic/lib/migration/plugins/plugin2"
 )
